pkg/web: move http definition to a separate package
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 7cc18852-5245-47e1-a119-d35364ed3156
workspace_id: c8c3ba95-0d12-4938-b5e1-cfa3447464dc
